Create a folder for buffers module
`SendBuffer` is about to receive a major overhaul, so it requires
its own space. It is better to make this change in advance so subsequent
reviews are easier to follow.

There are no functional changes in this PR, only refactoring
